requirements.txt
discord.py
mcrcon
